# src/main/java/com/dux/api_soccer_team/model/SoccerTeamModel.java
package com.dux.api_soccer_team.model;


import jakarta.persistence.*;
import jakarta.validation.constraints.NotNull;
import lombok.*;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
@Entity
@Table(name = "SoccerTeam")
public class SoccerTeamModel {
    @Id
    @GeneratedValue(strategy= GenerationType.IDENTITY)
    private Integer id;
    @Column(nullable = false)
    private String nombre;
    @Column(nullable = false)
    private String liga;
    @Column(nullable = false)
    private String pais;
}